Use stars for bullets, but there's some issues with fonts.
This should be changed to use an SVG instead, but I'm lazy right now. In Firefox
and Chrome on macOS, I found that U+2605 would not respect font-family and would
always default to the system font. This means that there is serious potential
for variation between the star appearances. It should be fixed!